Move interface Node to codeCharta.model #471
