+++ visualization/app/codeCharta/codeCharta.model.ts
@@ -1,4 +1,4 @@
-import { Vector3 } from "three"
+import {Vector3} from "three"
 
 export interface NameDataPair {
     fileName: string
@@ -172,4 +172,27 @@ export type RecursivePartial<T> = {
 
 export interface KeyValuePair {
     [key: string]: number;
+}
+
+export interface Node {
+    name: string;
+    width: number;
+    height: number;
+    length: number;
+    depth: number;
+    x0: number;
+    z0: number;
+    y0: number;
+    isLeaf: boolean;
+    deltas: KeyValuePair;
+    attributes: KeyValuePair;
+    children: Node[];
+    parent: Node;
+    heightDelta: number;
+    visible: boolean;
+    path: string;
+    origin: string;
+    link: string;
+    markingColor: string;
+    flat: boolean;
 }
