+++ visualization/app/codeCharta/ui/mapTreeView/mapTreeView.level.component.spec.ts
@@ -7,7 +7,7 @@ import { CodeMapHelper } from "../../util/codeMapHelper"
 import { IRootScopeService } from "angular"
 import { instantiateModule, getService } from "../../../../mocks/ng.mockhelper"
 import { CodeMapBuilding } from "../codeMap/rendering/codeMapBuilding"
-import { Node } from "../codeMap/rendering/node"
+import { Node } from "../../codeCharta.model"
 import { CodeMapBuildingTransition } from "../codeMap/codeMap.mouseEvent.service"
 import { ThreeOrbitControlsService } from "../codeMap/threeViewer/threeOrbitControlsService"
 import { CodeMapNode, BlacklistType, MarkedPackage } from "../../codeCharta.model"
